Day 6 Log Added
I updated the nav pills to be appropriately sized rather than full-width. I added centered and styled tags to the details of each work. I also added a newsletter sign-up and linked it with my mailchimp account. I plan to use this list to send a monthly summary email of my blog posts.