Create 1_Characteristics
On Day 1 of the 100 Days of Code challenge I wrote a simple console application to ask the users name and to tell him/her about some characteristics of the Python programming language. When the user types in the name of one of the keywords, the definition of that keyword appears.