Update formatting & Add template suggestion
Hi! I've found it difficult to understand that this file was supposed to be templates for people joining the challenge, so I edited it to make sure people could understand that better. Besides that, I added a template that I've been using since day one. Hope this helps people along the way, and thank you for the impact you've made in the dev community.

Cheers,
Victor.